requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
redis==5.2.0
sqlparse==0.5.1
